order-service/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0

sqlalchemy==2.0.30
pymysql==1.1.0

grpcio==1.64.0
grpcio-tools==1.64.0

python-dotenv==1.0.1
cryptography==42.0.5

# RabbitMQ
aio-pika==9.0.3
pika==1.3.2

pydantic_settings